+++ lib/src/widgets/embeds.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import '../../extensions.dart';
 import '../models/documents/nodes/leaf.dart' as leaf;
 import '../models/themes/quill_dialog_theme.dart';
 import '../models/themes/quill_icon_theme.dart';
@@ -15,12 +16,15 @@ abstract class EmbedBuilder {
     return WidgetSpan(child: widget);
   }
 
+  String toPlainText(Embed node) => Embed.kObjectReplacementCharacter;
+
   Widget build(
     BuildContext context,
     QuillController controller,
     leaf.Embed node,
     bool readOnly,
     bool inline,
+    TextStyle textStyle,
   );
 }