+++ lib/src/widgets/text_block.dart
@@ -152,7 +152,7 @@ class EditableTextBlock extends StatelessWidget {
             onLaunchUrl: onLaunchUrl,
             customLinkPrefixes: customLinkPrefixes,
           ),
-          _getIndentWidth(),
+          _getIndentWidth(context, count),
           _getSpacingForLine(line, index, count, defaultStyles),
           textDirection,
           textSelection,
@@ -168,47 +168,64 @@ class EditableTextBlock extends StatelessWidget {
     return children.toList(growable: false);
   }
 
+  double _numberPointWidth(double fontSize, int count) {
+    final length = '$count'.length;
+    switch (length) {
+      case 1:
+      case 2:
+        return fontSize * 2;
+      default:
+        // 3 -> 2.5
+        // 4 -> 3
+        // 5 -> 3.5
+        return fontSize * (length - (length - 2) / 2);
+    }
+  }
+
   Widget? _buildLeading(BuildContext context, Line line, int index,
       Map<int, int> indentLevelCounts, int count) {
-    final defaultStyles = QuillStyles.getStyles(context, false);
+    final defaultStyles = QuillStyles.getStyles(context, false)!;
+    final fontSize = defaultStyles.paragraph?.style.fontSize ?? 16;
     final attrs = line.style.attributes;
+
     if (attrs[Attribute.list.key] == Attribute.ol) {
       return QuillNumberPoint(
         index: index,
         indentLevelCounts: indentLevelCounts,
         count: count,
-        style: defaultStyles!.leading!.style,
+        style: defaultStyles.leading!.style,
         attrs: attrs,
-        width: 32,
-        padding: 8,
+        width: _numberPointWidth(fontSize, count),
+        padding: fontSize / 2,
       );
     }
 
     if (attrs[Attribute.list.key] == Attribute.ul) {
       return QuillBulletPoint(
         style:
-            defaultStyles!.leading!.style.copyWith(fontWeight: FontWeight.bold),
-        width: 32,
+            defaultStyles.leading!.style.copyWith(fontWeight: FontWeight.bold),
+        width: fontSize * 2,
+        padding: fontSize / 2,
       );
     }
 
     if (attrs[Attribute.list.key] == Attribute.checked) {
       return CheckboxPoint(
-        size: 14,
+        size: fontSize,
         value: true,
         enabled: !readOnly,
         onChanged: (checked) => onCheckboxTap(line.documentOffset, checked),
-        uiBuilder: defaultStyles?.lists?.checkboxUIBuilder,
+        uiBuilder: defaultStyles.lists?.checkboxUIBuilder,
       );
     }
 
     if (attrs[Attribute.list.key] == Attribute.unchecked) {
       return CheckboxPoint(
-        size: 14,
+        size: fontSize,
         value: false,
         enabled: !readOnly,
         onChanged: (checked) => onCheckboxTap(line.documentOffset, checked),
-        uiBuilder: defaultStyles?.lists?.checkboxUIBuilder,
+        uiBuilder: defaultStyles.lists?.checkboxUIBuilder,
       );
     }
 
@@ -217,35 +234,41 @@ class EditableTextBlock extends StatelessWidget {
         index: index,
         indentLevelCounts: indentLevelCounts,
         count: count,
-        style: defaultStyles!.code!.style
+        style: defaultStyles.code!.style
             .copyWith(color: defaultStyles.code!.style.color!.withOpacity(0.4)),
-        width: 32,
+        width: _numberPointWidth(fontSize, count),
         attrs: attrs,
-        padding: 16,
+        padding: fontSize,
         withDot: false,
       );
     }
     return null;
   }
 
-  double _getIndentWidth() {
+  double _getIndentWidth(BuildContext context, int count) {
+    final defaultStyles = QuillStyles.getStyles(context, false)!;
+    final fontSize = defaultStyles.paragraph?.style.fontSize ?? 16;
     final attrs = block.style.attributes;
 
     final indent = attrs[Attribute.indent.key];
     var extraIndent = 0.0;
     if (indent != null && indent.value != null) {
-      extraIndent = 16.0 * indent.value;
+      extraIndent = fontSize * indent.value;
     }
 
     if (attrs.containsKey(Attribute.blockQuote.key)) {
-      return 16.0 + extraIndent;
+      return fontSize + extraIndent;
     }
 
     var baseIndent = 0.0;
 
-    if (attrs.containsKey(Attribute.list.key) ||
-        attrs.containsKey(Attribute.codeBlock.key)) {
-      baseIndent = 32.0;
+    if (attrs.containsKey(Attribute.list.key)) {
+      baseIndent = fontSize * 2;
+      if (attrs[Attribute.list.key] == Attribute.ol) {
+        baseIndent = _numberPointWidth(fontSize, count);
+      } else if (attrs.containsKey(Attribute.codeBlock.key)) {
+        baseIndent = _numberPointWidth(fontSize, count);
+      }
     }
 
     return baseIndent + extraIndent;